authentification/requirements.txt
Flask
python-dotenv
bcrypt
psycopg2-binary
